+++ src/main/java/com/virtukch/dongiveupbe/domain/quiz_comment/dto/QuizCommentResponseDto.java
@@ -1,6 +1,6 @@
-package com.virtukch.dongiveupbe.domain.comment.dto;
+package com.virtukch.dongiveupbe.domain.quiz_comment.dto;
 
-import com.virtukch.dongiveupbe.domain.comment.entity.QuizComment;
+import com.virtukch.dongiveupbe.domain.quiz_comment.entity.QuizComment;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
